+++ src/apps/api/auth/auth.controller.js
@@ -68,22 +68,6 @@ export async function postLogin(req, res) {
 
   logger.info(`User id ${user.id} has logged-in!`);
 
-  const agent = useragent.parse(req.headers['user-agent']).toString();
-
-  let onlineUsers = await redis.get('online-users');
-
-  if (onlineUsers === null) {
-    await redis.set('online-users', JSON.stringify([]));
-    onlineUsers = await redis.get('online-users');
-  }
-
-  onlineUsers = JSON.parse(onlineUsers);
-  onlineUsers.push({ ...user, agent });
-  await redis.set('online-users', JSON.stringify(onlineUsers));
-
-  // req.io.sockets.emit('online-user', onlineUsers);
-  req.io.emit('online-user', onlineUsers);
-
   res.status(StatusCodes.OK).json({
     status: 'success',
     request_url: req.originalUrl,
@@ -130,11 +114,12 @@ export async function postSignup(req, res) {
   const { verify } = req.query;
   if (verify === true) {
     const date = new Date();
-    const [verifiedUser] = await AuthQueries.verifyUser(user.id, date);
+
+    await AuthQueries.verifyUser(user.id, date);
 
     logger.info(`User id ${user.id} was auto verified because of verify query!`);
 
-    const gde = await generateDefaultExercises(user.id);
+    generateDefaultExercises(user.id);
 
     logger.info(`Generated default exercises for User id ${user.id}!`);
 
@@ -159,7 +144,7 @@ export async function postSignup(req, res) {
   }
 
   // send verification email
-  await EmailService.send({
+  EmailService.send({
     to: newUser.email,
     subject: 'Verify Email',
     template: 'verify-email',
@@ -194,7 +179,7 @@ export async function getVerifyEmail(req, res) {
 
   logger.info(`User id ${uid} was successfully verified!`);
 
-  const gde = await generateDefaultExercises(uid);
+  generateDefaultExercises(uid);
 
   logger.info(`Generated default exercises for User id ${uid}!`);
 
@@ -239,7 +224,7 @@ export async function getReverify(req, res) {
       }
 
       // re send verification email
-      await EmailService.send({
+      EmailService.send({
         to: user.email,
         subject: 'Verify Email',
         template: 'verify-email',
@@ -306,7 +291,7 @@ export async function postForgetPassword(req, res) {
       throw new CustomError.BadRequestError(msg);
     }
 
-    const sent = await EmailService.send({
+    EmailService.send({
       to: user.email,
       subject: 'Password Reset',
       template: 'forget-password',
@@ -317,8 +302,6 @@ export async function postForgetPassword(req, res) {
     });
 
     logger.info(`Password reset link was send to ${email}`);
-
-    if (!sent) throw new CustomError.BadRequestError(`Something went went wrong while sending password reset to ${user.email}!`); // prettier-ignore
   }
 
   // but we send this regardless