+++ src/apps/api/contact/contact.controller.js
@@ -2,6 +2,8 @@ import EmailService from '../../../services/email.service.js';
 import logger from '../../../utils/logger.js';
 import { StatusCodes } from 'http-status-codes';
 import { email as envEmail } from '../../../config/env.js';
+import axios from 'axios';
+import { GITHUB } from '../../../config/env.js';
 
 /**
  * It sends an email to the user with the data they submitted
@@ -10,7 +12,32 @@ import { email as envEmail } from '../../../config/env.js';
  */
 export async function postContact(req, res) {
   const { email, subject } = req.body;
-  await EmailService.send({
+
+  if (subject === 'BUG' || subject === 'FEATURE') {
+    const data = {
+      title: subject,
+      body: req.body.message,
+      assignees: ['wajeht'],
+      labels: [subject.toLowerCase()],
+    };
+
+    const config = {
+      headers: {
+        Accept: 'application/vnd.github+json',
+        Authorization: `Bearer ${GITHUB.api_key}`,
+        'X-GitHub-Api-Version': '2022-11-28',
+      },
+    };
+
+    axios
+      .post(GITHUB.issue_url, data, config)
+      .then((response) => {})
+      .catch((error) => {
+        logger.error(`Something went wrong while sending the issue to GitHub: ${error}`);
+      });
+  }
+
+  EmailService.send({
     to: envEmail.auth_email,
     subject,
     template: 'contact',