+++ src/apps/app.js
@@ -20,7 +20,11 @@ import { red } from '../utils/rainbow-log.js';
 
 const app = express();
 const server = http.createServer(app);
-const io = new Server(server);
+const io = new Server(server, {
+  cors: {
+    origin: '*',
+  },
+});
 
 app.use((req, res, next) => {
   res.setHeader('X-Powered-By', 'Caddy, Express, Apache, Nginx, Nuxt, IIS');
@@ -42,27 +46,63 @@ app.use(express.urlencoded({ extended: true }));
 app.use(cookieParser(jwt_secret));
 app.use(
   express.static(path.resolve(path.join(process.cwd(), 'src', 'public')), {
-    maxage: '24h',
+    // 30 days in miliseconds
+    maxage: 2592000000,
   }),
 );
 
 app.use('/docs/*', (req, res, next) => Middlewares.authenticateUser(req, res, next, true));
 expressJSDocSwagger(app)(expressJsdocOptions);
 
-io.on('connection', function (socket) {
+io.on('connection', async function (socket) {
   logger.info(`socket.io connection was made!, ${socket.id}`);
 
-  // TODO: refactor this!!
-  socket.on('logout-user', async (user_id) => {
-    let users = await redis.get('online-users');
-    users = JSON.parse(users);
-    users = users.filter((u) => u.id != user_id);
-    await redis.set('online-users', JSON.stringify(users));
-    io.emit('online-user', users);
+  let onlineUsers = JSON.parse(await redis.get('onlineUsers')) || [];
+
+  socket.emit('onlineUser', onlineUsers);
+
+  socket.on('onlineUser', async (userData) => {
+    logger.info(`onlineUser event was fired!, ${socket.id}`);
+
+    const userExist = onlineUsers.some(
+      (user) => user.id === userData.id && socket.id === user.socket_id,
+    );
+
+    if (!userExist) {
+      onlineUsers.push({
+        ...userData,
+        socket_id: socket.id,
+      });
+
+      await redis.set('onlineUsers', JSON.stringify(onlineUsers));
+      socket.broadcast.emit('onlineUser', onlineUsers);
+    }
   });
 
-  socket.on('disconnect', () => {
+  socket.on('userDisconnected', async (userData) => {
+    logger.info(`userDisconnected event was fired!, ${socket.id}`);
+    let onlineUsers = JSON.parse(await redis.get('onlineUsers')) || [];
+
+    onlineUsers = onlineUsers.filter(
+      (user) => user.id !== userData.id && user.socket_id !== socket.id,
+    );
+
+    await redis.set('onlineUsers', JSON.stringify(onlineUsers));
+
+    socket.broadcast.emit('userDisconnected', userData.id);
+  });
+
+  socket.on('disconnect', async () => {
     logger.info(`socket.io connection was dropped!, ${socket.id}`);
+
+    let onlineUsers = JSON.parse(await redis.get('onlineUsers')) || [];
+
+    onlineUsers = onlineUsers.filter((user) => user.socket_id !== socket.id);
+
+    await redis.set('onlineUsers', JSON.stringify(onlineUsers));
+
+    // Broadcast the disconnected user's ID to all clients
+    socket.broadcast.emit('userDisconnected', socket.id);
   });
 });
 
@@ -89,7 +129,6 @@ app.use((req, res, next) => {
   // matching /api/v[number]/
   const isApiPrefix = req.url.match(/\/api\/v\d\//g);
 
-  // console.log(req.url);
   if (isApiPrefix) {
     throw new CustomError.BadRequestError('The resource does not exist!');
   }