+++ src/apps/ui/components/dashboard/LogoutButton.vue
@@ -5,9 +5,7 @@ import { useRouter, useRoute } from 'vue-router';
 import { sleep, isMobile } from '../../../../utils/helpers.js';
 import api from '../../../../utils/fetch-with-style.js';
 import useUserStore from '../../store/user.store.js';
-import { io } from 'socket.io-client';
 
-const socket = io('/');
 const router = useRouter();
 const userStore = useUserStore();
 const loading = ref(false);
@@ -15,8 +13,6 @@ const loading = ref(false);
 async function logout() {
   try {
     loading.value = true;
-
-    await sleep(800);
 
     const res = await api.get('/api/auth/logout');
     const json = await res.json();
@@ -30,8 +26,6 @@ async function logout() {
       }
     }
 
-    socket.emit('logout-user', userStore.user.id);
-
     userStore.isLoggedIn = false;
     userStore.clearUserInfo();
 
@@ -39,6 +33,13 @@ async function logout() {
     if (isMobile()) {
       logoutLink = '/dashboard/login';
     }
+
+    const userWithSocketId = {
+      ...userStore.user,
+      socket_id: window.socket.id,
+    };
+
+    window.socket.emit('userDisconnected', userWithSocketId);
 
     router.push({ path: logoutLink });
   } catch (e) {
@@ -55,7 +56,7 @@ async function logout() {
 </script>
 
 <template>
-  <button @click="logout()" class="btn btn-sm btn-danger" :disabled="loading">
+  <button @click="logout()" class="btn btn-danger" :disabled="loading">
     <div v-if="loading" class="spinner-border spinner-border-sm" role="status">
       <span class="visually-hidden">Loading...</span>
     </div>